+++ packages/database/prisma/seed.ts
@@ -13,7 +13,7 @@ async function main() {
 
 main()
   .catch((e) => {
-    console.error(e)
+    log(e)
     process.exit(1)
   })
   .finally(async () => {
